requirements.txt
python-telegram-bot==20.8
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
schedule==1.2.0
lxml==5.1.0 